Fixed warning in pr-31

+++ admin/class-memberpress-discord-admin.php
@@ -581,6 +581,7 @@ class Memberpress_Discord_Admin {
 		$ets_memberpress_discord_send_membership_expired_dm = sanitize_text_field( trim( get_option( 'ets_memberpress_discord_send_membership_expired_dm' ) ) );
 		$ets_memberpress_discord_send_membership_cancel_dm  = sanitize_text_field( trim( get_option( 'ets_memberpress_discord_send_membership_cancel_dm' ) ) );
 		$access_token                                       = get_user_meta( $user_id, '_ets_memberpress_discord_access_token', true );
+    $curr_level_id = NULL;
 		if ( ! empty( $access_token ) ) {
 			if ( $expired_membership ) {
 				$curr_level_id = $expired_membership['product_id'];
@@ -589,12 +590,16 @@ class Memberpress_Discord_Admin {
 				$curr_level_id = $cancelled_membership['product_id'];
 			}
 
-			$_ets_memberpress_discord_role_id = sanitize_text_field( trim( get_user_meta( $user_id, '_ets_memberpress_discord_role_id_for_' . $curr_level_id, true ) ) );
-			// delete already assigned role.
-			if ( isset( $_ets_memberpress_discord_role_id ) && $_ets_memberpress_discord_role_id != '' && $_ets_memberpress_discord_role_id != 'none' ) {
-					$this->memberpress_delete_discord_role( $user_id, $_ets_memberpress_discord_role_id, $is_schedule );
-					delete_user_meta( $user_id, '_ets_memberpress_discord_role_id_for_' . $curr_level_id, true );
+			if( $curr_level_id!==NULL ) {
+				$_ets_memberpress_discord_role_id = sanitize_text_field( trim( get_user_meta( $user_id, '_ets_memberpress_discord_role_id_for_' . $curr_level_id, true ) ) );
+				// delete already assigned role.
+				if ( isset( $_ets_memberpress_discord_role_id ) && $_ets_memberpress_discord_role_id != '' && $_ets_memberpress_discord_role_id != 'none' ) {
+						$this->memberpress_delete_discord_role( $user_id, $_ets_memberpress_discord_role_id, $is_schedule );
+						delete_user_meta( $user_id, '_ets_memberpress_discord_role_id_for_' . $curr_level_id, true );
+				}
+				delete_user_meta( $user_id, '_ets_memberpress_discord_expitration_warning_dm_for_' . $curr_level_id );
 			}
+
 			if ( is_array( $active_memberships ) && count( $active_memberships ) != 0 ) {
 				// Assign role which is mapped to the mmebership level.
 				foreach ( $active_memberships as $active_membership ) {
@@ -603,6 +608,7 @@ class Memberpress_Discord_Admin {
 						if ( $mapped_role_id && $expired_membership == false && $cancelled_membership == false ) {
 							$plugin_public->put_discord_role_api( $user_id, $mapped_role_id, $is_schedule );
 							update_user_meta( $user_id, '_ets_memberpress_discord_role_id_for_' . $active_membership->product_id, $mapped_role_id );
+              delete_user_meta( $user_id, '_ets_memberpress_discord_expitration_warning_dm_for_' . $active_membership->product_id );
 						}
 					}
 				}
@@ -625,7 +631,6 @@ class Memberpress_Discord_Admin {
 			if ( isset( $user_id ) && $allow_none_member == 'no' && empty( $active_memberships ) ) {
 				$plugin_public->memberpress_delete_member_from_guild( $user_id, false );
 			}
-			delete_user_meta( $user_id, '_ets_memberpress_discord_expitration_warning_dm_for_' . $curr_level_id );
 
 			// Send DM about expiry, but only when allow_none_member setting is yes
 			if ( $ets_memberpress_discord_send_membership_expired_dm == true && $expired_membership !== false && $allow_none_member = 'yes' ) {
